+++ test/python/qobj/test_qobj.py
@@ -18,7 +18,6 @@ from qiskit.qobj.exceptions import SchemaValidationError
 from qiskit.qobj import Qobj, QobjConfig, QobjExperiment, QobjInstruction
 from qiskit.qobj import QobjHeader, validate_qobj_against_schema
 from qiskit.providers.builtinsimulators import simulatorsjob
-from qiskit.providers.ibmq import ibmqjob
 from qiskit.test import QiskitTestCase
 from qiskit.test.mock import FakeRueschlikon
 
@@ -105,17 +104,6 @@ class TestQobj(QiskitTestCase):
 
         with self.assertRaises(SchemaValidationError):
             job = simulatorsjob.SimulatorsJob(backend, job_id, _nop, self.bad_qobj)
-            job.submit()
-
-    def test_ibmqobj_raises_error_when_sending_bad_qobj(self):
-        """Test IBMQobj is denied resource request access when given an invalid Qobj instance."""
-
-        backend = FakeRueschlikon()
-        self.bad_qobj.header = QobjHeader(backend_name=backend.name())
-
-        api_stub = {}
-        with self.assertRaises(SchemaValidationError):
-            job = ibmqjob.IBMQJob(backend, None, api_stub, 'True', self.bad_qobj)
             job.submit()
 
     def test_change_qobj_after_compile(self):
